# optimize bloom filter implementation