--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,8 @@
+module github.com/segmentio/bloom
+
+go 1.13
+
+require (
+	github.com/spaolacci/murmur3 v1.1.0
+	github.com/willf/bitset v1.1.10
+)