Hello, and thanks for this package!

This PR modifies the bloom filter implementation to avoid recreating the hasher for every operation, which was the only cause of heap allocations in this code. I believe this is OK since the documentation states that the BloomFilter type isn't safe to use concurrently from multiple goroutines.

Here are results from the benchmarks in this package before and after the change:
```
$ benchcmp /tmp/bench{1,2}
benchmark                       old ns/op     new ns/op     delta
BenchmarkEstimated              0.12          0.08          -29.74%
BenchmarkSeparateTestAndAdd     410           395           -3.66%
BenchmarkCombinedTestAndAdd     385           349           -9.35%

benchmark                       old allocs     new allocs     delta
BenchmarkEstimated              0              0              +0.00%
BenchmarkSeparateTestAndAdd     4              0              -100.00%
BenchmarkCombinedTestAndAdd     2              0              -100.00%

benchmark                       old bytes     new bytes     delta
BenchmarkEstimated              0             0             +0.00%
BenchmarkSeparateTestAndAdd     194           0             -100.00%
BenchmarkCombinedTestAndAdd     97            0             -100.00%
```

Let me know if you have any concerns about merging this change!