@lemire 2021-04-29
This was targeting a real problem, but with PR https://github.com/bits-and-blooms/bloom/pull/57 merged, I believe it is obsolete. If you disagree, please create a new PR.